+++ src/worker-app-state.mjs
@@ -1,0 +1,7 @@
+import { Store } from './store.mjs'
+
+globalThis.ficusjs = globalThis.ficusjs || {}
+
+globalThis.ficusjs.createAppState = globalThis.ficusjs.createAppState || function (options) {
+  return new Store(options)
+}