+++ test/e2e/app-with-worker-app-state/clear-button.mjs
@@ -1,0 +1,13 @@
+import { html, createCustomElement, withWorkerStore } from '../util/component.mjs'
+import { worker } from './worker-store.mjs'
+
+createCustomElement('clear-button',
+  withWorkerStore(worker, {
+    clear () {
+      this.dispatch('clear')
+    },
+    render () {
+      return html`<button type="button" onclick=${this.clear}>Clear</button>`
+    }
+  })
+)