+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2021-07-17
+
+### New
+- Add iife build for creating app state in web workers - `worker-app-state.iife.js`
+- Add `withWorkerStore` function for using stores created in web workers
+
 ## [1.1.1] - 2021-07-10
 
 ### Fixes